requirements.txt
easydict==1.9
gibson2==1.0.2
gym==0.17.2
imutils==0.5.3
joblib==1.0.1
matplotlib==3.3.1
natsort==8.1.0
numpy==1.19.1
open3d==0.13.0
opencv_python==4.4.0.42
pandas==1.1.5
Pillow==9.2.0
psutil==5.7.2
pudb==2019.2
pyglet==1.5.0
PyYAML==6.0
scikit_fmm==2021.2.2
scikit_learn==1.1.1
scipy==1.5.2
seaborn==0.11.2
setuptools==49.6.0.post20200814
six==1.15.0
spinup==0.1
tensorflow==2.9.1
torch==1.9.0
torchvision==0.10.0

-e extensions/gym-foo
-e extensions/rlpyt